refactor(trace): separate TeXRA-specific helpers from AgentTrace surface
Splits the trace API into two layers so SDK consumers see only
agent-general events while TeXRA-internal code keeps its sugar:

  src/agent/trace/  (general)         src/logger/  (TeXRA)
  ─────────────────────────────       ─────────────────────────
  AgentEvent union                    TexraTrace extends AgentTrace
  AgentTrace interface                TexraTraceEmitter extends
  TraceEmitter (general only)           TraceEmitter
  noopTrace                           noopTexraTrace

AgentEvent shed `files.loaded` (TeXRA-shaped FileListEntry payload) and
the strict MessageType enum on LogEvent (now generic `string`). Those
concerns live in TexraTraceEmitter, which emits filesLoaded as a
`domain` event and TeXRA log lines with MESSAGE_TYPES strings the
transcript recorder still understands.

What moved off AgentTrace (now only on TexraTrace):
- logError / logProgress / logErrorData
- logInternal / debugInternal / logScratchpad
- logContextManagement / logContextState
- missingOutputs / latexDiff / userMessage
- filesLoaded / logFileCategory
- emitToolUse / logToolUseStart / updateToolUse
- logWebSearch / logWebFetch / statistics
- startGroup / endGroup

What stayed on AgentTrace (lean, agent-general):
- emit / subscribe / activeStageId / withStage
- debug / info / warn / error (with optional messageType: string)
- openStage / openStream
- usage / contextState / toolStart / toolEnd / domain

`RunContext.trace` and `AgentCore.logger` are typed as `TexraTrace`
because the runtime context is TeXRA-bound. SDK consumers can still
program against `AgentTrace` since `TexraTrace extends AgentTrace`.

Also added docs/architecture/agent-trace.md with ASCII diagrams of
the two-layer interface, file layout, and single-resolve invariant.

Verified: typecheck + test-kernel + lint all clean; 1042/1042 tests pass.
