+++ src/agent/modelHandlers/anthropicContextManagement.ts
@@ -1,5 +1,5 @@
 // Type imports - agent
-import type { AgentTrace } from '@agent/trace';
+import type { TexraTrace } from '@logger';
 
 // Type imports - Anthropic SDK
 import type { AnthropicBeta } from '@anthropic-ai/sdk/resources/beta/beta';
@@ -162,7 +162,7 @@ export function setupContextManagement(
 export function logContextManagementFromResponse(
   response: BetaMessage,
   contextWindow: number,
-  logger: AgentTrace,
+  logger: TexraTrace,
 ): void {
   const totalInputTokens =
     response.usage.input_tokens +
