+++ src/agent/output/LatexDiffManager.ts
@@ -1,13 +1,14 @@
 import * as path from 'path';
 
 import { platform } from '@platform/platform';
-import type { AgentTrace, AgentLogStage } from '@agent/trace';
+import type { AgentLogStage } from '@agent/trace';
 import { AgentWorkflowSetting } from '@agent/core/AgentDataclass';
 import { getWorkspaceState } from '@agent/core/stateStore';
 import { toErrorMessage } from '@common/errors';
 import { WorkspaceStateKey } from '@common/state/stateKeys';
 import { compileLatex2Pdf } from '@latex/texTools';
 import { LaTeXdiffResult, LaTeXdiffService } from '@latex/latexdiff';
+import type { TexraTrace } from '@logger';
 import {
   type DiffResult,
   type ExecutionId,
@@ -48,7 +49,7 @@ export class LatexDiffManager {
     private readonly agentSetting: AgentWorkflowSetting,
     private readonly getOutputFiles: () => { [key: number]: OutputFileInfo[] },
     private readonly baseFiles: FileLocation[],
-    private readonly logger: AgentTrace,
+    private readonly logger: TexraTrace,
     private readonly streamId: string,
     private readonly fileService: TaskRunFileService,
   ) {
