+++ src/agent/output/compileCheck.ts
@@ -1,11 +1,11 @@
 import * as path from 'path';
 
-import type { AgentTrace } from '@agent/trace';
 import { getWorkspaceState } from '@agent/core/stateStore';
 import { toErrorMessage } from '@common/errors';
 import { WorkspaceStateKey } from '@common/state/stateKeys';
 import { hasLatexCompiler } from '@latex/latexToolchain';
 import { compileLatex2Pdf } from '@latex/texTools';
+import type { TexraTrace } from '@logger';
 import type { CompileFailure, OutputFileInfo } from '@shared/schemas';
 import {
   LATEX_CONFIG_DEFAULTS,
@@ -28,7 +28,7 @@ import { getOutputFilesByRound, type OutputState } from './outputState';
 export interface CompileCheckContext {
   fileService: TaskRunFileService;
   outputState: OutputState;
-  logger: AgentTrace;
+  logger: TexraTrace;
   streamId: string;
 }
