+++ src/test-kernel/agent/modelHandlers/OpenAIResponseCompaction.vitest.ts
@@ -7,13 +7,13 @@ import {
 } from 'llm-zoo';
 
 // Local imports - agent model handlers
-import type { AgentTrace } from '@agent/trace';
 import { ModelHandlerOpenAIResponse } from '@agent/modelHandlers/modelHandlerOpenAIResponse';
+import type { TexraTrace } from '@logger';
 
 // Type imports
 import type { ResponseInputItem } from 'openai/resources/responses/responses';
 
-function createLoggerStub(): Partial<AgentTrace> & { streamId: string } {
+function createLoggerStub(): Partial<TexraTrace> & { streamId: string } {
   return {
     streamId: 'test-channel',
     debug: vi.fn(),
@@ -48,7 +48,7 @@ function createConfig(overrides: Partial<ModelConfig> = {}): ModelConfig {
 
 function createHandler(): ModelHandlerOpenAIResponse {
   const handler = new ModelHandlerOpenAIResponse(createConfig());
-  handler.setLogger(createLoggerStub() as unknown as AgentTrace);
+  handler.setLogger(createLoggerStub() as unknown as TexraTrace);
   (handler as { getStreamingConfig: () => boolean }).getStreamingConfig = () =>
     false;
   return handler;
