+++ src/test-kernel/agent/output/OutputProgressEvents.vitest.ts
@@ -2,7 +2,6 @@
 import { describe, expect, it } from 'vitest';
 
 // Local imports
-import type { AgentTrace } from '@agent/trace';
 import { AgentCategory } from '@agent/core/AgentDataclass';
 import { OutputNode } from '@agent/implementations/flows/reflection/nodes/OutputNode';
 import type { ReflectionServices } from '@agent/implementations/flows/reflection/ReflectionServices';
@@ -13,6 +12,7 @@ import {
 } from '@agent/output/OutputFileProcessor';
 import type { XmlOutputManager } from '@agent/output/XmlOutputManager';
 import { normalizeRunId } from '@common/constants/runIds';
+import type { TexraTrace } from '@logger';
 import type {
   AgentFileLocation,
   FileLocation,
@@ -108,7 +108,7 @@ describe('output progress events', () => {
     const logger = {
       debug: () => {},
       missingOutputs: () => {},
-    } as unknown as AgentTrace;
+    } as unknown as TexraTrace;
     const xmlManager = {
       processSingleXmlOutput: async () => {
         throw new Error('invalid xml');
