+++ src/test/agent/modelHandlers/ModelHandlerAnthropic.test.ts
@@ -15,11 +15,11 @@ import {
   ModelProvider,
   ReasoningEffort,
 } from 'llm-zoo';
-import type { AgentTrace } from '@agent/trace';
 import type { AgentConfig } from '@agent/core/AgentConfig';
 import { AgentCategory, AgentSettingSchema } from '@agent/core/AgentDataclass';
 import { AgentWorkspaceState } from '@agent/core/AgentWorkspaceState';
 import { ModelHandlerAnthropic } from '@agent/modelHandlers/modelHandlerAnthropic';
+import type { TexraTrace } from '@logger';
 
 // Type imports
 
@@ -107,7 +107,7 @@ function stubHandlerForTest(
   handler: ModelHandlerAnthropic,
   loggerOverrides?: Partial<Record<string, unknown>>,
 ): void {
-  handler.setLogger(createLoggerStub(loggerOverrides) as unknown as AgentTrace);
+  handler.setLogger(createLoggerStub(loggerOverrides) as unknown as TexraTrace);
   (handler as any).getStreamingConfig = () => false;
 }
 
@@ -1246,7 +1246,7 @@ describe('ModelHandlerAnthropic message guards', () => {
         logContextManagement: (message: string, data: unknown) => {
           events.push({ message, data });
         },
-      }) as unknown as AgentTrace,
+      }) as unknown as TexraTrace,
     );
 
     (handler as any).logContextManagementFromResponse(
@@ -1521,7 +1521,7 @@ describe('ModelHandlerAnthropic pre-message_start error handling', () => {
       supportsTokenCounting: false,
       supportsReasoning: false,
     });
-    handler.setLogger(createLoggerStub() as unknown as AgentTrace);
+    handler.setLogger(createLoggerStub() as unknown as TexraTrace);
     // Force streaming path so we exercise the pre-message_start catch block.
     (handler as any).getStreamingConfig = () => true;
