+++ src/test/agent/modelHandlers/ModelHandlerOpenAINormalize.test.ts
@@ -9,16 +9,16 @@ import {
   type ModelConfig,
   ModelProvider,
 } from 'llm-zoo';
-import type { AgentTrace } from '@agent/trace';
 import { ModelHandlerDashScope } from '@agent/modelHandlers/modelHandlerDashScope';
 import { ModelHandlerDeepSeek } from '@agent/modelHandlers/modelHandlerDeepSeek';
 import { ModelHandlerKimi } from '@agent/modelHandlers/modelHandlerKimi';
+import type { TexraTrace } from '@logger';
 
 // Type imports
 
 // Local imports - model config
 
-type LoggerStub = Partial<AgentTrace> & {
+type LoggerStub = Partial<TexraTrace> & {
   streamId: string;
   debugMessages: string[];
   infoMessages: string[];
@@ -129,7 +129,7 @@ describe('ModelHandlerOpenAI.normalizeMessages hook', () => {
     });
     const handler = new ModelHandlerDeepSeek(config);
     const loggerStub = createLoggerStub();
-    handler.setLogger(loggerStub as unknown as AgentTrace);
+    handler.setLogger(loggerStub as unknown as TexraTrace);
     (handler as any).getStreamingConfig = () => false;
 
     const client = createClientStub();
@@ -169,7 +169,7 @@ describe('ModelHandlerOpenAI.normalizeMessages hook', () => {
     });
     const handler = new ModelHandlerKimi(config);
     const loggerStub = createLoggerStub();
-    handler.setLogger(loggerStub as unknown as AgentTrace);
+    handler.setLogger(loggerStub as unknown as TexraTrace);
     (handler as any).getStreamingConfig = () => false;
 
     const client = createClientStub();
@@ -195,7 +195,7 @@ describe('ModelHandlerOpenAI.normalizeMessages hook', () => {
     });
     const handler = new ModelHandlerDashScope(config);
     const loggerStub = createLoggerStub();
-    handler.setLogger(loggerStub as unknown as AgentTrace);
+    handler.setLogger(loggerStub as unknown as TexraTrace);
     (handler as any).getStreamingConfig = () => false;
 
     const client = createClientStub();
